data_stream  top level -> stream
