+++ oceanics-io-wasm-ui/src/stream/mod.rs
@@ -1,0 +1,8 @@
+pub mod interactive_data_stream;
+pub mod style;
+pub mod data_stream;
+
+/**
+ * Enable plotting 2D data series to a canvas.
+ */
+pub mod stream {}
