+++ oceanics-io-wasm-ui/src/webgl/texture_2d.rs
@@ -31,6 +31,4 @@ pub mod texture_2d {
             ).unwrap();
         }
     }
-
-    
 }
